Fix handling of linux bond interfaces
__NOTE__

This patch set contains a static reference to a in-flight
charm-helpers review. The c-h change must be merged, released
and the static reference removed before merging this patch set.

__NOTE__

At present linux bonds will be ignored when mentioned directly
in configuration options parsed by the ``BridgePortInterface``
data structure.

If a mac address of a physical interface making up the bond is
listed the code correctly finds the bond name and adds that to
the bridge.

Virtual interfaces are ignored when then map is built, linux bonds
are virtual interfaces.

This patch fixes the problem by checking physical interfaces for
bond membership while building the map. If an interface is member
of a bond add the bond name to the map.

Previously the bond membership check was only done when looking
up interface name from mac after building the map.

Change-Id: I09b17cc039c622373f05423fe9993e0254145f28
Closes-Bug: #1883244