Add configuration option for Neutron OVN Metadata workers
The Neutron default of number of CPUs / 2 is way too high for the
Neutron OVN Metadata service and puts an unneccessary toll on the
OVN Southbound database.

Add a new configuration option to control this which defaults to
2.

Note that we do not use the default worker-multiplier handling
code that many other of our charms use because the main use case
for that code is to handle the use case where API services run in
containers.

The Neutron OVN Metadata service runs directly on the host.

Merge after openstack-charmers/charm-layer-ovn#21

Change-Id: Ic335a552dc9afcaab026c534499e4070846851eb
Closes-Bug: #1893656
(cherry picked from commit e06b578b4580c0f46fe38f8937934a6ff72b97a3)