[WIP] SR-IOV: match on PCI address, don't do runtime config
Populate the sriov-netplan-shim interfaces.yaml config with PCI
address of interfaces. This is necessary as the VFs may be
configured at a time in the boot sequence where interface naming
has not settled yet.

Stop doing runtime reconfiguration of SR-IOV, in addition to being
detrimental to any virtual machine instance consuming the VF this
will break NIC firmware in some configurations.

Merge after:
https://github.com/juju/charm-helpers/pull/549
https://github.com/openstack-charmers/charm-layer-ovn/pull/31

Change-Id: I397a2a82501c565724d6fce6b49f7b68d57d47fd
Closes-Bug: #1908351