Provide chassis certificates to principal
This is useful for related charms providing payloads that
expect direct communication with the OVN databases.

While a principal and subordinate charm executes in the same
environment, the payload usually execute under different service
accounts and as such it is impractical and may be less secure to
attempt to provide direct on-disk file access.

Related-Bug: #1918271
Change-Id: I3ded62818a43a05c67903882f8bf0e32e23e7e07