Update patch set 1
Patch Set 1:

(1 comment)

The build is breaking due to the layer addition as a git subproject; the build system doesn't know what to do with it.  We don't have a good story on how to use the CI to test layers; typically we do it in a bastion, so this does show up an issue.  However, an alternative method would be to use a lock file with just the layer in it, and you can specify all parts of the build as you wish.  Take a look at the charm-build --help command, or contact me directly.

Patch-set: 1
Reviewer: Gerrit User 20870 <20870@4a232e18-c5a9-48ee-94c0-e04e7cca6543>