Update patch set 2
Patch Set 2:

Gabriel, now that the ovn-layer change has merged, you're good to go to repropose this change without cargoing the layer code. You can put "Func-Test-PR:" in the commit message to get the new tests be run with the change without changing src/test-requirements.txt.

Chances are you will end up with an empty change, but that is still required for validation and to trigger a new charm publication.

When there are no other changes you can generate an UUID and replace the uuid in the file called `rebuild`.

Patch-set: 2
CC: Gerrit User 13686 <13686@4a232e18-c5a9-48ee-94c0-e04e7cca6543>