Update patch set 4
Patch Set 4:

Looking at the job output the new test did not run, and I see that is because of me not being clear on how you make it run. After the "Func-Test-PR:" in the commit message you will need to put in the https link to the GitHub PR. In essence if you replace "zaza-openstack-tests PR:" with "Func-Test-PR:" and keep the URL behind in your commit message your tests will run.

Patch-set: 4