Update to build using charmcraft - 20.03 batch
Batch to add building by charmcraft for the ovn charms for the 20.03
batch.

build.lock: Update commit for layer-ovn

Fix UCA source in bundles

At present the bundles will not test what is on the tin.

The ovn-chassis charm is a subordinate charm and requires its
principle charm to configure the correct UCA pocket on the
unit.  The magpie charm has grown a ``source`` configuration
option for this purpose.

Change-Id: I1d9e9c6b733cb66c518b0b505a796ddb6879553e