Update to build using charmcraft - 20.12 batch
Batch to add building by charmcraft for the ovn charms for the 20.12
batch.

build.lock: Update commit for layer-ovn

Fix UCA source in bundles

At present the bundles will not test what is on the tin.

The ovn-chassis charm is a subordinate charm and requires its
principle charm to configure the correct UCA pocket on the
unit.  The magpie charm has grown a ``source`` configuration
option for this purpose.

Change-Id: I396e296c968419d1106ca76b68605fce527ec681