Automatic setup/teardown of iptables NOTRACK rules for excluding GENEVE traffic from nf_conntrack
On busy cloud environments, the number of unique GENEVE flows that
goes through ovn-chassis could be high. This is due to fact that GENEVE
uses source port randomization to achieve an even distribution across
multiple links. This randomization causes nf_conntrack table to filled with
many GENEVE flows, eventually leading to connection drops in a busy environment.
As there is no particular reason for tracking these flows, the solution for this
problem is to simply exclude GENEVE traffic from nf_conntrack tracking.
This can be done via putting rules with `-j NOTRACK` jump into relevant iptables
chains, which many people are already using as a solution to this problem.
The problem with that solution is iptables rules are not persisted across reboots
and the rules are not tied to charm's presence or status.

This enhancement introduces a configurable feature to automatically setup and teardown
a NOTRACK rule chain. The feature is controlled via `enable-notrack-rules` config
flag and if enabled, a rule chain with name `juju-och-nfnotrack` will be created
and the chain will be populated with rules using the criteria list specified in
`notrack-rule-criteria`(*) configuration value. The resulting chain will be attached
to `PREROUTING` and `OUTPUT` chains in the `raw` table.

(*) `notrack-rule-criteria` is by default set to a rule criteria that excludes GENEVE
traffic.

Closes-bug: #1978806
Change-Id: Ieb47c0aebaad43e15c6a10f6e9ca3fcc5619dacc