Add bindep for CI Python dependencies and pin tox
The opendev CI environment has recently transitioned to using
Ubuntu Jammy as the base OS for test runs.

A side effect of this transition is a change in some of the
dependencies for lint and unit tests.

tox.ini: Remove charm-tools pin as it pulls in obsolete
dependencies + pin tox < 4.0.0.

Change-Id: I9d52fa71ce98f6e77462d6f8ab52b62fd787c356