Build on 23.04 base to run on 23.04 (Lunar)
Currently the charm is built on 22.04 which has Python 3.10, so when
building the python binary wheels they are associated to Python 3.10,
this makes them unsable on Ubuntu 23.04 which has Python 3.11.

This change makes the building process for Lunar as a separate job.

Closes-Bug: #2019996
Change-Id: Ib4146880f59bd4f8b1d26efa0d84f798e71268ab
(cherry picked from commit 9d55de7bc3ca470dee5a06dfcf66131aa9174cd8)