NO MERGE: Test using binary wheels for charm build
This may improve our build time in CI which currently hovers
around 15 mins. Perhaps by opting to NOT build our dependencies
from source, we could improve that time.

Even if this change is effective, we will have to ensure that
the option is used only in dev/ci environments and not when
building charm for publishing.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>